Consolidate worklist filters views/omponents into their own directory
