+++ src/js/i18n/en-US.yml
@@ -616,17 +616,18 @@ careOptsFrontend:
           bodyText: Are you sure you want to delete the selected Flows? This cannot be undone.
           headingText: Delete Flows?
           submitText: Delete
-      dateFilterViews:
-        actionViews:
-          month: This Month
-          today: Today
-          yesterday: Yesterday
-        added: 'Added:'
-        defaultDateView:
-          thisMonth: This Month
-        relativeDateView:
-          today: Today
-          yesterday: Yesterday
+      filters:
+        dateFilterViews:
+          actionViews:
+            month: This Month
+            today: Today
+            yesterday: Yesterday
+          added: 'Added:'
+          defaultDateView:
+            thisMonth: This Month
+          relativeDateView:
+            today: Today
+            yesterday: Yesterday
       filtersApp:
         allGroupsName: All Groups
         ownerFilterHeadingText: Filter by Owner
