+++ src/js/views/patients/worklist/filters/date-filter_views.js
@@ -8,7 +8,7 @@ const DateView = View.extend({
   className: 'button-filter js-date',
   template: hbs`
     {{far "calendar-alt"}}{{~ remove_whitespace ~}}
-    {{@intl.patients.worklist.dateFilterViews.added}} {{formatDateTime selectedDate "MM/DD/YYYY"}}
+    {{@intl.patients.worklist.filters.dateFilterViews.added}} {{formatDateTime selectedDate "MM/DD/YYYY"}}
   `,
   triggers: {
     'click': 'click',
@@ -20,7 +20,7 @@ const MonthView = View.extend({
   className: 'button-filter js-date',
   template: hbs`
     {{far "calendar-alt"}}{{~ remove_whitespace ~}}
-    {{ @intl.patients.worklist.dateFilterViews.added }} {{formatDateTime selectedMonth "MMM YYYY"}}
+    {{ @intl.patients.worklist.filters.dateFilterViews.added }} {{formatDateTime selectedMonth "MMM YYYY"}}
   `,
   triggers: {
     'click': 'click',
@@ -32,11 +32,11 @@ const RelativeDateView = View.extend({
   className: 'button-filter js-date',
   template: hbs`
     {{far "calendar-alt"}}{{~ remove_whitespace ~}}
-    {{ @intl.patients.worklist.dateFilterViews.added }} {{date}}
+    {{ @intl.patients.worklist.filters.dateFilterViews.added }} {{date}}
   `,
   templateContext() {
     return {
-      date: intl.patients.worklist.dateFilterViews.relativeDateView[this.model.get('relativeDate')],
+      date: intl.patients.worklist.filters.dateFilterViews.relativeDateView[this.model.get('relativeDate')],
     };
   },
   triggers: {
@@ -49,7 +49,7 @@ const DefaultDateView = View.extend({
   className: 'button-filter js-date',
   template: hbs`
     {{far "calendar-alt"}}{{~ remove_whitespace ~}}
-    {{ @intl.patients.worklist.dateFilterViews.added}} {{ @intl.patients.worklist.dateFilterViews.defaultDateView.thisMonth }}
+    {{ @intl.patients.worklist.filters.dateFilterViews.added}} {{ @intl.patients.worklist.filters.dateFilterViews.defaultDateView.thisMonth }}
   `,
   triggers: {
     'click': 'click',
@@ -58,9 +58,9 @@ const DefaultDateView = View.extend({
 
 const ActionsView = View.extend({
   template: hbs`
-    <button class="datepicker__button js-yesterday">{{ @intl.patients.worklist.dateFilterViews.actionViews.yesterday }}</button>{{~ remove_whitespace ~}}
-    <button class="datepicker__button js-today">{{ @intl.patients.worklist.dateFilterViews.actionViews.today }}</button>{{~ remove_whitespace ~}}
-    <button class="datepicker__button js-current-month">{{ @intl.patients.worklist.dateFilterViews.actionViews.month }}</button>
+    <button class="datepicker__button js-yesterday">{{ @intl.patients.worklist.filters.dateFilterViews.actionViews.yesterday }}</button>{{~ remove_whitespace ~}}
+    <button class="datepicker__button js-today">{{ @intl.patients.worklist.filters.dateFilterViews.actionViews.today }}</button>{{~ remove_whitespace ~}}
+    <button class="datepicker__button js-current-month">{{ @intl.patients.worklist.filters.dateFilterViews.actionViews.month }}</button>
   `,
   triggers: {
     'click .js-yesterday': 'click:yesterday',
