--- src/js/views/patients/worklist/role-component.scss
@@ -1,19 +1,0 @@
-.role-component__role {
-  color: $black-60;
-  font-size: 12px;
-  margin-left: 4px;
-}
-
-.is-highlighted .role-component__role {
-  color: $white;
-}
-
-.role-component__button {
-  .fa-angle-down {
-    margin-left: 8px;
-  }
-
-  .fa-user-circle {
-    margin-right: 8px;
-  }
-}
